Change in captcha validation
