--- code/TrainningVivek/PopupLogin/etc/config.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0"?>
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:module:Magento_Store:etc/config.xsd">
-    <default>
-        <customer>
-            <captcha>
-                <always_for>
-                    <custom_form>1</custom_form>
-                </always_for>
-            </captcha>
-        </customer>
-        <captcha translate="label">
-            <frontend>
-                <areas>
-                    <custom_form>
-                        <label>Custom Form</label>
-                    </custom_form>
-                </areas>
-            </frontend>
-        </captcha>
-    </default>
-</config>
