+++ code/TrainningVivek/PopupLogin/etc/di.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0"?>
+<!--
+/**
+* Copyright © Magento, Inc. All rights reserved.
+* See COPYING.txt for license details.
+*/
+-->
+<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
+    <type name="TrainningVivek\PopupLogin\Controller\Customer\Ajax\Register">
+        <plugin name="captcha_ajax" type="TrainningVivek\PopupLogin\Plugin\AjaxCaptchaValidation" sortOrder="50" />
+    </type>
+</config>
