--- code/TrainningVivek/PopupLogin/etc/frontend/events.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0"?>
-<!--
-/**
- * Copyright © Magento, Inc. All rights reserved.
- * See COPYING.txt for license details.
- */
--->
-<config xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:Event/etc/events.xsd">
-    <event name="controller_action_predispatch_training_customer_ajax_register">
-        <observer name="captcha" instance="TrainningVivek\PopupLogin\Observer\CheckRegisterObserver" />
-    </event>
-</config>
